chore: moved hrm types to own package hrm-types
