+++ hrm-domain/hrm-service/tsconfig.build.json
@@ -15,6 +15,7 @@
     { "path": "packages/service-discovery" },
     { "path": "resources-domain/packages/resources-search-config" },
     { "path": "resources-domain/resources-service" },
+    { "path": "hrm-domain/packages/hrm-types" },
     { "path": "hrm-domain/packages/hrm-search-config" },
     { "path": "hrm-domain/hrm-core" },
     { "path": "hrm-domain/scheduled-tasks/hrm-data-pull" },
