+++ hrm-domain/packages/hrm-types/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["*.ts"],
+  "compilerOptions": {
+    "outDir": "./dist"
+  }
+}
+
