+++ tsconfig.json
@@ -17,6 +17,7 @@
     { "path": "packages/service-discovery" },
     { "path": "lambdas/packages/hrm-authentication" },
     { "path": "lambdas/job-complete" },
+    { "path": "hrm-domain/packages/hrm-types" },
     { "path": "hrm-domain/packages/hrm-search-config" },
     { "path": "hrm-domain/lambdas/contact-complete" },
     { "path": "hrm-domain/lambdas/contact-retrieve-transcript" },
